refactoring; update beautifyStackTrace to take all types of errors
