+++ lib/utils/index.js
@@ -4,7 +4,6 @@ const mkpath = require('mkpath');
 const glob = require('glob');
 const lodashMerge = require('lodash.merge');
 const {By} = require('selenium-webdriver');
-const stackTraceParser = require('stacktrace-parser'); 
 
 const Logger = require('./logger');
 const BrowserName = require('./browsername.js');
@@ -19,7 +18,7 @@ const requireModule = require('./requireModule.js');
 const getAllClassMethodNames = require('./getAllClassMethodNames.js');
 const printVersionInfo = require('./printVersionInfo.js');
 const {filterStack, filterStackTrace, showStackTrace, stackTraceFilter, errorToStackTrace} = require('./stackTrace.js');
-
+const beautifyStackTrace = require('./beautifyStackTrace');
 const formatRegExp = /%[sdj%]/g;
 const testSuiteNameRegxp = /(_|-|\.)*([A-Z]*)/g;
 const nameSeparatorRegxp = /(\s|\/)/;
@@ -538,30 +537,6 @@ class Utils {
   static shouldReplaceStack(err) {
     return !Utils.alwaysDisplayError(err);
   }
-
-
-  /**
-   * Read the User file from the stackTrace and forms 
-   * string with highlighting the line with error.
-   */ 
-  static beautifyStackTrace(err) {
-    if (err && ['NightwatchAssertError'].includes(err.name)) {
-      const parsedStack = stackTraceParser.parse(filterStackTrace(err.stack))[0];
-      let lines = require('fs').readFileSync(parsedStack.file, 'utf-8').split(/\r?\n/).reduce(function(lines, newLine, lineIndex) {
-        const currentLine = lineIndex + 1;
-        if (currentLine === parsedStack.lineNumber) {
-          lines += Logger.colors.cyan(`\t${currentLine} ${newLine}`);
-        } else if (currentLine <= (parsedStack.lineNumber + 2) && currentLine >= (parsedStack.lineNumber - 2)) {
-          lines +=  (`\t${currentLine} ${newLine}\n`);
-        }
-
-        return lines;
-      }, '');
-      lines += '\n';
-
-      return lines;
-    }
-  }
 }
 
 lodashMerge(Utils, {
@@ -585,7 +560,9 @@ lodashMerge(Utils, {
   TimedCallback,
 
   getFreePort,
-  printVersionInfo
+  printVersionInfo,
+
+  beautifyStackTrace
 });
 
 module.exports = Utils;
